--- a/betfairlightweight/__version__.py
+++ b/betfairlightweight/__version__.py
@@ -1,6 +1,6 @@
 __title__ = "betfairlightweight"
 __description__ = "Lightweight python wrapper for Betfair API-NG"
 __url__ = "https://github.com/liampauling/betfair"
-__version__ = "2.10.0"
+__version__ = "2.10.1"
 __author__ = "Liam Pauling"
 __license__ = "MIT"