--- a/requirements.txt
+++ b/requirements.txt
@@ -1,1 +1,1 @@
-requests<2.25.0
+requests<2.26.0