--- a/tests/resources/streaming_ocm_EMPTY_IMAGE.json
+++ b/tests/resources/streaming_ocm_EMPTY_IMAGE.json
@@ -1,0 +1,1 @@
+{"op":"ocm","id":2,"clk":"AAgAAAAAAAAAAA==","pt":1604418055629,"oc":[{"accountId":1234567,"fullImage":true,"id":"1.161613698"}]}