…ice, once on on_process() and once on snap()

Refactor to only call during on_process if there is queue
Timings:
  before: 44734 calls    1.381s    0.000   48.707s
  after: 22367 calls    0.752s    0.000   26.373s

empty image test added